feat: add current-display value reporting (current day, zone, time skew) and setting of zone

+++ webserver.go
@@ -114,7 +114,7 @@ func webserver(port int) {
 			log.Printf("invalid zone numner")
 		} else {
 			params := TStatZoneParams{}
-			flags := byte(0)
+			flags := uint16(0)
 			zi := zn - 1
 
 			if len(args.FanMode) > 0 {
